[IMP] web_studio: drag and drop to add new elements in views
We want to avoid unnecessary clicks to add new elements in the form view.
This is why we implemented drag & drop.
In addition, this makes it possible to unify the behavior between the
Wesite Editor and Studio.

In the new "add" tab in the sidebar, it is now possible to add:
- components: notebook, group (only for form view).
- Existing fields: only those who are not already in the view.
- New Fields: char, text, integer, float, html,...
- Widgets: image, tags. (it just a new field but the widget is already on it)

For some new fiels (selection and relational fields), we have to display
a dialog to ask mandatory datas.
In others cases, the element is created automaticaly.

For this moment, we only provide this functionality for the form and list view
but in the future we will add this feature in all concerned views.
So, the "add" tab is temporarily locked.

Special thanks to @sri for the design.