[IMP] website_*: sort and filter for portal document
Make the portal module use the generic 'sort and filter'
mecanism.

website_contract already implement the feature, but now it uses
the generic template.

website_helpdesk did not use pager, archive and sorting
features. It would have been sad to not provide these
feature, since it is a new module.