[IMP] mrp_plm: take into account 'Consumed in operation' on BoM line.
When calculating the BoM changes field on the ECO, we did not take into
account the field on the BoM line which indicates during which operation
the product is consumed.  Now, we do together with the UoM of the product.
That way, a change of UoM, will be shown as an Update instead of an Add / Remove.

In order to make it clear which field changed, when a field is updated,
it is shown as old_value -> new_value.