[IMP] web_studio: enable navigation in Studio
Before this commit, it was not possible to navigate while in Studio.
You needed to leave Studio, navigate and then go back in Studio again.

The navigation is now possible and the URLs are correctly handled.

Notice that you will get an error if you open an action not editable by Studio.