[FIX] web_studio: Show invisible elements properly
Two issues are fixed by this commit:

(1) When you clicked on "Show invisible elements" in "view" tab,
you were automatically redirected to the "properties" tab.
Old state informations was send to the new form_editor.
Now, we correctly reset the node clicked when you switch tabs.

(2) The checked value of the input named "show_invisible" was not
saved. So, when you checked this input, the sidebar was rerendered
and the checked value was lost.
Now, the field is saved in sidebar.