[IMP] web_studio: edit field description if label edited
As the field name is created randomly with Studio (x_new_field_...),
just changing the label in the view is not enough because one cannot
be able to find it in another view.

This commit edits the field description (for custom fields) when the user
edits the field label with Studio.