[IMP] stock_barcode: scan product packaging
When you scan the barcode of a product packaging in a picking it adds
the quantity that contains the the product packaging to the
corresponding pack operation.

When you scan the barcode of a product packaging during an inventory, it
adds the quantity that contains the the product packaging to the
corresponfing inventory line.