[ADD] delivery_barcode
When a barcode is set on a delivery packaging type we are able to scan
it in the picking, and get the "put in pack" action triggered with the
corresponding delivery packaging type.

We have added a computed field on picking to handle in javascript the
delivery packaging available for a carrier.