[FIX] sale_ebay: Remove settings link to a new module website_sale_ebay
We want to display on the website res config settings a link to configure ebay when the ecommerce is installed

We do not have other choices than creating a brige module between website_sale and sale_ebay and move this view into it.

We don't want to create a bridge module for only one small thing like that.