[IMP] project_timesheet_synchro: adapt with ts sheet removal
Since the hr_timesheet_sheet is removed, we 'hack' the
data format returned to the app timesheet to keep
compatibility without chaging anything in the JS code.
The value of the sheet state is harcoded for the app :
it does not change anything for the display at the end,
so the state is 'open'.

This commit also update the test suite.