[FIX] web_enterprise, web_studio: handle not-square app icons
In studio, you can upload an image to use for an app in the
appswitcher. If the uploaded image is not a square, then the
appswitcher could appear broken if the app is the last one on
the row.

This commit changes the way images are handled by the appswitcher,
again. Now all app icons are a div.o_app_icon element. In the case
the app has an image, then this image is used as background-image
of this div (placed at the center and filling the square div).
In the case the app has an icon (studio), then the div contains
a <i/> element which displays the icon, positioned at the center
of the div.