[IMP] web_studio: always allow to add custom attrs on form fields
Before this commit, the user had the possibility to tell if a form
field was required, readonly and/or invisible. In the case, a field
had a custom attrs coming from (demo) data, the corresponding
attribute(s) were marked as "custom" and the custom button allowed
to *see* the associated domain.

Now, the checkbox for each of those three attributes is always
associated with the custom button which now allows to *edit* the
associated domain. To do that, the DomainSelector had to be adapted
to only display the fields which are already in the view and to
prevent from following relations.
(see commit https://github.com/odoo/odoo/commit/495b5ef1172ff9b477c2ebf9afb9ad8d2ca2e989)

Note: this commit also fixes a "bug" from the previous version:
when clicking on the checkbox when a custom attrs is set, the
custom attrs is removed and the checkbox is thus unchecked (before,
the attrs stayed so unchecking the checkbox had no visual effect,
but double clicking on it fully checked it (as the attribute was then
forced to true and ignored the attrs even if it was still there)).