[FIX] web_studio: more specific xpath for kanban edition
A <field> can be either placed at the beginning of the <kanban> template,
before the <templates> (this is useful when field.raw_value is needed) or
inside the <templates> (use to display the field in the view).

We thus need to change the xpath from //field[...] to //templates//field[...]
when targetting a field in kanban.