[FIX] web_studio: new one2many creation
When creating a new one2many field, one has to choose a many2one that
points to the current model ; from this selection, it should not
be possible to create an new many2one field for another model directly.

One should previously create a many2one field and select it afterwards.