[FIX] web_studio: do not push state for Studio actions
Since the navigation inside Studio has been activated, it is now possible
to switch to another action while in Studio. The problem occurs when
we switch to another action which shouldn't be consider as a 'normal'
action (i.e. we don't want to edit this action with Studio). When
refreshing the page, we want to be in Studio to edit this first action.

To do so, the option `keep_state` has been introduced in the action
description (and not in options because the `do_push_state` only have
the action).