[FIX] web_studio: show fields with no label and empty value in form
In form, if a field node is not in a group node, the label is not shown.
So, nothing appears if there is no value for this field.
Example: You don't see anything if you drag a new field outsite of a group.

To show these fields, we show the display name as default value and we apply a different style
to make the difference with others fields (same behavior as in Kanban view).