[FIX] web_studio: display empty widget
In the kanban editor, if a field is not set, we display
it anyway with its string for value.

If a widget is set on this field but this widget does
not support a string for value (ex: the monetary widget
only supports numbers), it's a bad idea to use the widget renderer
(that will crash) so we just set the widget text.

Also, the value 0 should not be considered as empty.