[FIX] web_studio: sidebar: filter widgets based on field types
In the sidebar, when you click on a field, you are able to set a widget for this field.
But all widgets appears and some of them are not adapted to the type of field.

Now, when you create a widget, you are able to precise for what type of field it is intended.
In this way, we can filter the widget list from the sidebar.