[FIX] web_studio: navigate in account reports
The heuristic used to determine if we were navigating
inside Studio (and then open this action with Studio)
was based on the xml id, which is not exactly correct
(ex: account reports don't have a xml id).

The solution here is to use an option key `disable_edition`
which determine if the action needs to be open normally
and not edited with Studio.