[FIX] account_reports: fix followup summary generation + print followup pdf letter
summary was formatted but not returned, resulting in a message containing "dear %(partner_name)s" instead of "dear Agrolait". Also, summary should always be recomputed except when printing or sending by mail. If we don't do that, first time a followup will show, the text for first reminder will be saved and if we check later when the second reminder should be displayed, instead we will still find the first reminder letter
+ fix followup print letter which was broken due to commit 3d04e44880fb77812e65342ef25bb8ad06094ddc