[FIX] web_studio: edit the view after activating it
When the user activates a new view type, she can now directly
edit it without clicking a second time on the thumbnail. This
was causing some usability issues as the user didn't understand
the result of her action.

The tour has been adapted because we directly enter in the
kanban view at the end.