[REF] print: clean print wizard
Since it is not possible to have computed
field on the wizard line, we remove it.
They were not trigger since the wizard is not
created in db.
Use the default_get to generate line values, as
they are not meant to be modified (only removed).
This simplify the code and use the standart
mecanism.

Also move the complexity to determine if the
document is sendable on the document itself.