[IMP] print_docsaway: better currencies support
Docsaway supports now others currencies, but for different
purpose :
- get price : you set the currency on the request, and the price
is given in the specified currency
- get balance : the provider balance is returned expressed in
the currency of the account on docsaway.com
- send file : the returned price of the transaction is returned
in the currency of the docsaway.com account

The price and currency on the print.order makes no sense: the idea
is here to set the currency used by your docsaway.com account and
use it everywhere (for all call to docsaway API), even if it is
not the currency of the company. At least, the amount exposed
on form view will be set with their correct currency.

This should have been a fixed, but adding fields in stable version
is dangerous (and forbidden).