[FIX] web_studio: avoid adding monetary field on some models
We can add a monetary field only if there is a relational field to 'res.currency' to the model.
At this moment, this field has to be named 'currency_id'.

Before this commit, you were able to drag and drop a monetary field to any model.
So, the view was broken if there were no 'currency_id' on the model.

- The first problem is that we can't create a field named 'currency_id' with Studio because
  this field have to be named 'x currency_id' to keep changes after updating the concerned module.
  (A commit should follow to allow the name 'x_currency_id' in the ORM. @rco-odoo)

- The second problem is that we can't use the 'currency_field' attribute.
  This works only if you declare the field in Python.

Because of these two problems, we allow the addition of a monetary field only if we detect
the presence of a field named 'currency_id'.