[FIX] web_studio: remove flickering after old view edition
This commit only destroys the old view once the new is ready, as the
old views are not rendered synchronously.

Note that this reduces the flickering but doesn't remove it completely
for the calendar view as the lib doesn't allow the rendering in one pass.

This commit also removes the flickering on checkbox in the sidebar
when setting a field option.