[IMP] quality: removing unused SPC button and REF functions
We have removed the old button 'SPC' which does the same as the one
showing the average and standard deviation. And we also set an
ensure_one on the functions called by the buttons which cannot have
multiple records in self.

The method 'action_see_spc_control' doesn't use anymore the check that
have not been processed yet