[IMP] quality: quality checks and alerts on picking
First of all, we add a picking_id on quality alert to be able to link
and alert to the picking that has generated it.

We add a button on the picking view letting you generate some quality
alerts and one showing the number of alerts relating this picking.

We also made some cosmetic changes on the button showing the checks, to
make it green if all checks passed and red if there's at least one
failed