[FIX] web_studio: reloading the sidebar when adding a component
After this commit: c18d9a08, the sidebar wasn't reloaded when adding a new component.

There were multiple issues:
- An existing field no longer disappears from the sidebar when it is added to the view.
- There was also a side effect when you tried to add a new field;
  references to the previous operations wasn't reloaded and the result of this was that
  we had the same name for several new fields.

Now, we reload the sidebar only if we are not in the "Properties" tab to avoid flickering and
We have prevented multiple references to operations to avoid this side effect.

We also reset the scroll level after rendering in the sidebar.