[FIX] Plm: Problem with approvals
Use case when the problem happens:
* Create an engineering change order
* Set an approval on a stage
* Move the change order to that stage
* Reject the approval
* Move the change order backward in a stage without approvals
* Try to move the change order forward
This fail it display an error message asking you to approve the change
order before moving it even if you are on a stage without approvals.

The field allow_change_stage is the problem. This field prevent a change
order move if this one is not approved. It is set to False when
the approval is rejected, but it is never updated if you change the
order to a new stage.

This commit add stage_id to the depend property of the method that
update kanban state and allow_change_stage field. Now when you move
a change order, it write the new stage_id and trigger the function.