[FIX] web_studio: allow `Import` action
The action with target `new` (wizards) are blocked in Studio since
https://github.com/odoo/enterprise/commit/25351e47e3710c9ec7da57fe9ff2ef3929af748a

There is an exception though for the `Import` wizard in the app switcher
which can legitimately be opened in Studio.

This commit also removes `studio_on` in the web_client which was in fact
useless because `studio_mode` contains all the necessary information.