[FIX] web_studio: impossible to drop in 'autocompletion table'
When the div containing the three tables was not big enough to be behind
the 'autocompletion table', it was impossible to drop element inside
the 'autocompletion table'. The 'drop' event was then never triggered.
To fix it, we set the 'drop' event on each table and not in the div container
anymore.