[FIX] Plm: Reject change order stay rejected after some moves.
Use case when it happens:
*Move a change order to a stage with approvals
*Reject it
*Move it to a previous stage
*Move it to the stage where it was rejected
The change order state appears has rejected.

The problem come from approvals. Each time you move an order to a
new stage. It creates one approvals by approvals_templates of this
stage. When you reject a change order, it also rejects all the approvals
relative to this object. After some moves, if you put it in the stage
where it was rejected it creates new approvals for the objects but you
still have the olders that were rejected. A function check if there is
at least one rejected approval on the change order and if there is one
it display the change order has rejected.

This commit fix this bug by checking if there already exists one
approval by approval templates for the actual stage. If there exists,
it will updates them instead of create news. It will also add a note
in the chatter in order to keep a kind of history.

And comment only to comment approvals