[FIX] web_studio: show right fields for domain on x2x fields
On x2x fields you have the possibility to set a domain to only match
some of the records of the x2x relation when searching which one to
suggest during edition (e.g. on a crm.lead form view you have the
possibility to set tags (model crm.lead.tag); if you only want to
suggest the tags whose name begins by "o", you could use the domain
feature to do that).

In saas-14, a domain selector widget is instanciated instead of the
plain text edition of the domain. However, it was set up with the
wrong model, using the form view one instead of the relation one
(in the example above, it used crm.lead model instead of crm.lead.tag).
This induced a traceback each time trying to use a x2x field with a
custom studio-made domain.