[FIX] web_studio: ID should always be available for attrs domain
Since commit https://github.com/odoo/odoo/commit/68a97420c397902d65e363044440f7b112c20823,
the ID field can be used for form view attrs domains even if the ID
field is not explicitely added to the view. As said in mentioned
commit, the feature seemed to already work with list views.

In studio, when building an attrs domain, we only allow to use fields
which are already in the views. An exception had to be made to allow
the ID field too.