[IMP] quality_mrp: quality point on workcenter routing
When the settings 'Work Orders' are activated in the MRP module, the
user has the possibility to add some control points on the workcenters
that are raised when a workorder is done on this workcenter.

To give the possibility to users, to have different quality check on
the same workorder for different boms or products. The quality point are
now defined on a routing operatizon instead of a workcenter.