[FIX] web_studio: display empty widgets in kanban
Before this commit, the content of a  widget without value was replaced by
a span with its string.

By doing this, the widget was not clickable anymore and thus not editable.

Instead of replacing the whole content, we now set the widget text.

This commit also corrects the evaluation of the widget 'emptiness'.
The value of a relation widget is an array, which is evaluated to `true`
if empty.