[FIX] web_studio: prevent from adding binary in kanban
If a image (binary field) is added in the kanban view, the whole base64
will be displayed, which is not exactly what we want.
This is due to the fact that there is no widget image in the kanban view
registry (the kanban view is using a different field registry ; this will
be modified with the new views).

This commit thus prevents the user from adding this type of field in
the kanban view.