[IMP] account_report: add a 'partner ledger' button on res.partner form view
Its name switches to "Supplier Ledger" or "Customer Ledger" if the partner is *only* one of these. This button opens the partner ledger for the current partner.
When opened with a default_partner_id (hence: from a partner), the Partner Ledger now only displays the data related to this partner, unfolding it automatically for readabilty.