[IMP] web_studio: set the 'keep state' option at the right place
Before this commit, the action description was directly modified to set the
'keep state' option. This was not possible in all the cases, e.g.
the 'action_web_studio_form' action.

After this commit, the 'keep state' option can be set in the 'options'
argument and the 'do_action' funtcion will set it to the action.
This allow the behaviour expected in all the actions.