[FIX] web_studio: form_editor, minors UX
- Do not hide column labels on click:
Selecting a column the inner fields will loose opacity: it help understand what is selected (the column) and what’s not (the inner fields).
Since the previous implementation was too much generic, it impacted the column label too that should be clearly visible instead.

- 'Image Field', show available room for image file
The previous design was misleading because of the image size and colours.
This fix show clearly how much size will be available for the image file