[FIX] mrp_plm: versionning on product attachments
To avoid adding columns with a default value on a pottentially big table
we remove the override of the ir.attachments and so we do the
versionning by keeping a snapshot of the attachments on the ECOs.

So the revisions are visible through the ECO, and not directly through
the ir.attachments. We are also only show the attachments set on
product_template, has you can only set attachments through ECOs on
product_template and the attachments shown oby the module 'document' are
only the product_template ones.