[IMP] delivery_fedex: cache the rates returned by Fedex
This revision will make the rates asked to Fedex cached
for 4 hours for a same delivery service/weight/location/...

With this, if a same rate request is asked multiple times,
we avoid to ask Fedex multiple times for this same rate,
which didn't changed in a so short interval.

e.g. if a customer refreshes
the checkout page on the shop where he is asked to choose
the delivery service he would like, it will ask
Fedex only once the rate for his shipping, as long
as he do not change his order, and therefore the weight
of the shipping.

Courtesy of @beledouxdenis , slightly modified by myself.

opw-710963
opw-743133