[FIX] l10n_mx: fix bad references in demo data
Some demo data in l10n_mx are incorrect, preventing the installation
of this module on some demo/test databases.
The MX chart of account is not guaranteed to be loaded when l10n_mx_reports
is installed. To avoid this problem, the demo data is moved to a YAML
script that fallbacks to default values if the required MX COA data
is not available.

Closes odoo/enterprise#829