[FIX] mrp_plm: copy product_template attachments
When you start a new revision for product attachments, the attachments
are copied from the product to the ECO. And we try to unlink them from
the product template when an ECO is applied.

The ECO try to take the attachments from the field attachment_ids of the
product templat ewhich doesn't exist anymore since the rev: https://github.com/odoo/enterprise/commit/7e1c73cfc84cd23b6ba87dab713c3078de16568d

To fix this issue, we are using a search to take the arttachments of the
product template.

We also rename the computed field 'attachment_count' in
'template_attachment_count' because there is already a field
'attachment_count' defined in module 'website_sale_digital' which
doesn't match the same attachments.