[FIX] website_studio: website access rights
A website menu is created for the page containing the
newly created website form but this was causing an error if the user
is not in the website designer group. And you can be administrator,
have access to studio but not be in the website designer group.
This case is now managed with an error message.