[FIX] web_grid: new grid test was broken on firefox
Commit 94cff2577162b480f8da0a6ecf67b3fa9d47aa2d introduced a new test
for the grid view... but this test was buggy on firefox.
Unfortunately, as the test relies on the `window.getSelection` API,
this test has to be browser dependent.

Indeed:
* On Chrome, selecting the whole content of a node with an unique
text node thanks to the `Range.selectNodeContents` function
creates a Range element whose reference is the text node and the
selection end value is the text node length.
* On Firefox, doing the same operation creates a Range element
whose reference is the parent node and the selection end value
is 1 (saying that one whole text node is selected).