[FIX] web_studio: list editor forward-port
The commit 294ca6b has not been correctly forward-ported in e26acf0.

This code has nothing to do with the method `set_local_state` but
needs to be executed when adding a new field.

As the field drag & drop has been introduced in saas-14, the previous
method `on_new_column` was removed.

Complement to 6e21c4d.