[IMP] mrp_plm: Rebase ECO with other ECO and manual BoM changes
- In the case you manually change something on a BoM, it will add that
  change to the bom_rebase_ids of the active ECOs on that BoM.
- In the case you apply another ECO on the same BoM, it will set a
  field on your ECO to the BoM activated by the other ECO.  A computed field
  previous_change_ids will calculate the difference between your old
  BoM and the activated one as the changes to be applied on your new
  one.

The ECOs related to the changes will change to a 'rebase' state.  That
is when a Rebase button will appear on the ECO.  If the changes are
successfull, you can apply them.  If there are conflicts, the state will
change to 'conflict' and you are supposed to change them manually, after
which you can confirm it and apply the changes.

If the rebase is because of another ECO, it will set the old bom of your
ECO to the new bom of the other ECO.

Test cases are added with several ECOs on the same BoM with changes
directly on the BoM and several ECOs at the same time, with a
conflicting case and a normal one.