[FIX] mrp_plm: only set status comment on comment approvals
When a user comment an ECO, if there are any approvals of type 'comment'
exist for this user in the current stage, the status of those approvals
should be set at status 'comment'.

The current behavior is that when a user comment an ECO, all his
approvals on this stage are set to status 'comment'.

So to fix it, we only apply the status 'comment' on the approvals of
type 'comment' dor this user.