[FIX] web_studio: do not uselessly write on field
When modifying the string field in a view with Studio, we also modify
its field description.

This operation, which seems harmless, is quite costly (it requires
to reload the registry) so the less it's done the better.

But with the current behaviour, this operation is repeated each
time another one is applied.

We thus avoid to modify the field description if it's the same.