[IMP] web_studio: allow widget selection in kanban
The widget attribute selection was only enabled for the list and form view,
not for kanban. The main reason was because the kanban view was using another
field registry.

With the new views, the kanban view uses the same field registry ; we can thus
allow the user to change the widget attribute on a field.