[FIX] web_studio: allow to edit form view without currentId
When switching from a multi-records view to the form view in Studio,
there is no `currentId` set in the action environment. The user will thus
edit the form view without any record, even though there are existing ones.

We set here a default `currentId` when editing a view with Studio.