[FIX] web_studio: remove the right sidebar
When clicking on a node, the sidebar is updated, which means a new
sidebar is rendered and the old one is destroyed.

Before rendering the new sidebar, some rpcs could have been done which
caused a duplicated sidebar (the old one wasn't deleted).

We decided not to handle the case where the sidebar rendering is
asynchronous. This code will need to be adapted if the sidebar rendering
becomes asynchronous.