[IMP] mrp_maintenance: add groups/domain to workorder
Some items in view were not relevant if the mrp setting 'work order' was not activated.
In order to hide the previous items we added the group 'mrp.group_mrp_routings'
on relevant fields or button in the view.

For the domain on work orders, it checks if the work order's MO is the same than the MO selected.