[FIX] mrp_mps: bad use of this in an handler
Causing a traceback when opening Manufacturing > Reporting >
Master Production Schedule.

It would be nice to add a test (even very basic, just to ensure
that the client action can be rendered without crashing), but it
interacts with the action manager (its parent), e.g. to update
the control panel. This will be easier after the second part of
the webclient refactoring.