[FIX] mrp_mps: client action: listen to search view changes
When the user changes the filters in the search view, this one
triggers up an event to notify its environment that its state
changed. This came with the new views, as before the search view
triggered an event on itself (i.e. used 'trigger' instead of
'trigger_up'). At some point, 'trigger' will be deprecated and
'trigger_up' will be used instead, everywhere.

The code in mrp_mps hasn't been updated accordingly, so the view
wasn't updated when the user changed the filters.