[FIX] web_gantt: lazy loading of js libs
This commit fixes two issues related to the forward-port of commit
cf88aa673cb0c5dda735f774a164b272c30ebe8d.

First, it adds a fallback on the current locale to 'en_US',
because the code should not crash if this info is not available in
the user_context (for example, it is the case in the test
environment).

Second, the two js libs can't be loaded simultaneously (the one
added in rev. cf88aa673 must wait for the other one to be loaded).
So this rev. ensures that the libs are loaded sequentially.