[ADD] hr_contract_salary: Allow employees and applicant to configure their salary package
PURPOSE
=======

We want to allow employees and applicants to choose freely their wage and advantages according to the amount of money the employer is willing to give.

This configuration should be available on a frontend page, easy to use, on which the gross salary is updates automatically and on which the computation Gross -> Net Salary is also available.

SPECIFICATION
=============

Processes
~~~~~~~~~

A. An employee wants to modify his current contract
---------------------------------------------------

Go on the employee’s contract and, click on “Send contract modification”. The employee receives an email with a link to configure her salary package.

By clicking on this link, the employee can make changes on her current salary package. This won’t impact the cost for the employer. Once the changes are ok, the employee can click on “review & sign contract”. The employee will be asked to sign the new contract.  (*)

Note that the employee can also use this link to modify her personal information (address, person of contact in case of emergency, etc.).

It creates a new contract for the employee, in “new” stage. The HR manager can accept or refuse the request. (a) If the request is accepted, the contract is moved to the stage “running”, and the previous contract in the stage “running” is moved to the “expired” stage. (b) If the request is refused, the contract is moved to the stage “expired”.

(*) What happens if the employee quit the page before signing the contract ? The new contract is created in the backend and the HR manager can send back the eSign document.

B. Following an appraisal, an employee’s contract is modified
-------------------------------------------------------------

The HR manager can duplicate the current contract of the employee and modifying it (by changing the gross salary or the total cost for example). Then the HR manager can send the contract modification to the employee. The rest of the process doesn’t change.

C. An applicant receives an offer
---------------------------------

From an application, click on “Create Offer” to create a contract offer for this applicant. A new contract is created. You can adapt this contract if needed. Once the contract is ready, click on “Send Offer”. An email is sent to the applicant. It gives him an access to the salary package configurator.

The applicant can then adapt his contract. This won’t impact the cost for the employer. He has also to enter his personal information. Once the applicant has finished, he can click on “Review Contract & Sign”. He will be asked to sign his contract.

A new contract will be created in the backend. If the HR manager accepts it, the archived employee representing the applicant will be unarchived.

D. You make a demo of the salary package configurator
-----------------------------------------------------

If you want to give an access to the salary package configurator to someone who is neither an employee nor an applicant (e.g. for a demo), it’s possible.

Go on a job position and click on “configure package”. You can then play with the salary package configurator BUT you cannot sign and send (so there is no contract created in the backend).

Configuration
~~~~~~~~~~~~~

A. Document Template
--------------------

Create a Document Template in eSign.

This part is quite technical.

Go in Sign > Documents > Document Templates and upload a pdf template. Open this pdf template and drop fields (e.g. a text field) at places where some info has to be filled in.

If the information should be filled in automatically, enter in field name the technical name of the field from the hr.contract model. The responsible should be the HR responsible. Example : https://drive.google.com/file/d/0B0ITmdw6h3ZrbDlMVFBRX2d0Mmc/view.

B. Job Position
---------------

Create a Job Position.

On the Job Position, you define :
The default contract for new employees : that’s the standard offer that is usually made for this job position.
The new contract document template : the future employee signs this document when he accepts an offer.
The contract update document template : the employee signs this document when his contract is updated (in case of an appraisal for example).

PS: A huge thanks to Nicolas Michaux for all the design part and to Laura Piraux for all the specifications, functional reviews, ideas and followups.