[FIX] web_studio: destroy the correct editor after error
In Studio, when there is an error with the current operation (e.g. the
view can't be rendered), we undo this operation and render the previous
view.

In this case, the first editor wasn't correctly destroyed because the
second one was set as the current editor too soon, before being appended.

This commit also contains a small refactoring to put the `try catch` only
around the critical section.