[FIX] website_crm_score: don't use max lead scored to guess the last assignation.
Before this commit, we use the max lead_id with this score to find the last score assignation.
But in some cases (after a merge e.g.) we score manually a subselect of leads. So we cannot assume
that there are no other lead_id which need to be scored between the last automatic scoring and the
last leads merged that have been scored manually.