[IMP] website_calendar: display direct employee links
Use the direct employee link to book an appointment with a specific
employee.
Useful for employees to send out a personal appointment booking url.

This feature was already implemented but hidden.

This commit adds and applies a widget capable of displaying the url
in the employee list view present in the appointment type form view.