[FIX] quality : email alias does not create alert
Sending an email to the alias of a quality team doesn't create a
quality alert.

The problem is the attribute required for the field "product_tmpl_id" in
the model quality.alert. There is no way to guess the product template
id from an email. When the alias mail method try to create a quality
alert object, it returns an error because the product_tmpl_id is not
set.

This commit removes the required condition on product_tmpl_id.