[FIX] helpdesk: make test work again
Commit 38fbbbe0a fixed an issue with the sla computation cron which
added a new condition in the computation of a deadline: the sla
need to actually have changed for the new deadline to be computed.

This test was not compatible with this new condition since the
ticket is created with the correct sla coming directly from the
team. To actually recompute the deadline after setting the fake
creation date in the past, an explicit change of sla is needed.