[FIX] helpdesk: do not write on basicmodel record
The helpdesk module extends the kanban model and write on the data point
some extra information.  This was fine, until a recent commit in
community that optimise the result of the basicmodel get method.  In
that commit, instead of having a blacklist of keys that should not be
exported, we have a white list of keys that should be exported.

Therefore, we have to adapt this module to make it read information from
somewhere else.