[FIX] web_studio: show non empty x2many fields
Before this fix, the x2many widgets wasn't rendered correctly
if there is a value on it.
It was not possible to edit or remove this field.

In this version, the reimplemented views are not able to display
a default view. It must be specified in the arch view.

As we can't refactor the code in a stable branch, we only show the
name of the field instead of rendering the widget with the
corresponding records.

opw-741512