[FIX] web_studio: add a monetary field
Before adding a new monetary field, one should check if a currency field
exists on the model.

This was previously done with a RPC to search for such a field.
This RPC is not needed anymore with the new views as we have the information
about all current model fields in the `fieldsView`.